Rename WasmerRuntime to Wasmer2Runtime
