+++ examples/example-protocol/src/main.rs
@@ -304,8 +304,8 @@ fn main() {
                 .dependencies(PLUGIN_DEPENDENCIES.clone())
                 .build(),
         ),
-        BindingsType::RustWasmerRuntime,
-        BindingsType::RustWasmerWasiRuntime,
+        BindingsType::RustWasmer2Runtime,
+        BindingsType::RustWasmer2WasiRuntime,
         BindingsType::TsRuntime(
             TsRuntimeConfig::new()
                 .with_msgpack_module("https://unpkg.com/@msgpack/msgpack@2.7.2/mod.ts")
@@ -389,20 +389,20 @@ fn test_generate_rust_plugin_without_some_fields() {
 }
 
 #[test]
-fn test_generate_rust_wasmer_runtime() {
+fn test_generate_rust_wasmer2_runtime() {
     static FILES: &[(&str, &[u8])] = &[
         (
-            "bindings/rust-wasmer-runtime/bindings.rs",
-            include_bytes!("assets/rust_wasmer_runtime_test/expected_bindings.rs"),
+            "bindings/rust-wasmer2-runtime/bindings.rs",
+            include_bytes!("assets/rust_wasmer2_runtime_test/expected_bindings.rs"),
         ),
         (
-            "bindings/rust-wasmer-runtime/types.rs",
-            include_bytes!("assets/rust_wasmer_runtime_test/expected_types.rs"),
+            "bindings/rust-wasmer2-runtime/types.rs",
+            include_bytes!("assets/rust_wasmer2_runtime_test/expected_types.rs"),
         ),
     ];
     fp_bindgen!(BindingConfig {
-        bindings_type: BindingsType::RustWasmerRuntime,
-        path: "bindings/rust-wasmer-runtime",
+        bindings_type: BindingsType::RustWasmer2Runtime,
+        path: "bindings/rust-wasmer2-runtime",
     });
     for (path, expected) in FILES {
         tests::assert_file_eq(path, expected)
@@ -410,20 +410,20 @@ fn test_generate_rust_wasmer_runtime() {
 }
 
 #[test]
-fn test_generate_rust_wasmer_wasi_runtime() {
+fn test_generate_rust_wasmer2_wasi_runtime() {
     static FILES: &[(&str, &[u8])] = &[
         (
-            "bindings/rust-wasmer-wasi-runtime/bindings.rs",
-            include_bytes!("assets/rust_wasmer_wasi_runtime_test/expected_bindings.rs"),
+            "bindings/rust-wasmer2-wasi-runtime/bindings.rs",
+            include_bytes!("assets/rust_wasmer2_wasi_runtime_test/expected_bindings.rs"),
         ),
         (
-            "bindings/rust-wasmer-wasi-runtime/types.rs",
-            include_bytes!("assets/rust_wasmer_wasi_runtime_test/expected_types.rs"),
+            "bindings/rust-wasmer2-wasi-runtime/types.rs",
+            include_bytes!("assets/rust_wasmer2_wasi_runtime_test/expected_types.rs"),
         ),
     ];
     fp_bindgen!(BindingConfig {
-        bindings_type: BindingsType::RustWasmerWasiRuntime,
-        path: "bindings/rust-wasmer-wasi-runtime",
+        bindings_type: BindingsType::RustWasmer2WasiRuntime,
+        path: "bindings/rust-wasmer2-wasi-runtime",
     });
     for (path, expected) in FILES {
         tests::assert_file_eq(path, expected)
