--- examples/example-rust-wasmer-runtime/src/spec/types.rs
@@ -1,1 +1,0 @@
-../../../example-protocol/bindings/rust-wasmer-runtime/types.rs
