--- examples/example-rust-wasmer-runtime/src/wasi_spec/bindings.rs
@@ -1,1 +1,0 @@
-../../../example-protocol/bindings/rust-wasmer-wasi-runtime/bindings.rs
