--- examples/example-rust-wasmer-runtime/src/wasi_spec/types.rs
@@ -1,1 +1,0 @@
-../../../example-protocol/bindings/rust-wasmer-wasi-runtime/types.rs
