+++ examples/example-rust-wasmer2-runtime/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 edition = "2021"
-name = "example-rust-wasmer-runtime"
+name = "example-rust-wasmer2-runtime"
 version = "0.1.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
