+++ examples/example-rust-wasmer2-runtime/src/main.rs
