+++ examples/example-rust-wasmer2-runtime/src/spec/bindings.rs
@@ -1,0 +1,1 @@
+../../../example-protocol/bindings/rust-wasmer2-runtime/bindings.rs
