+++ examples/example-rust-wasmer2-runtime/src/spec/types.rs
@@ -1,0 +1,1 @@
+../../../example-protocol/bindings/rust-wasmer2-runtime/types.rs
