+++ examples/example-rust-wasmer2-runtime/src/test.rs
