+++ examples/example-rust-wasmer2-runtime/src/wasi_spec/bindings.rs
@@ -1,0 +1,1 @@
+../../../example-protocol/bindings/rust-wasmer2-wasi-runtime/bindings.rs
